Move runtime wiring into internal nest
Extract application setup and shutdown coordination from cmd/nest
so the command package can stay focused on process entrypoint
concerns.
